refactor(server): separate provider, models, and credentials into distinct domains
Move model registries (LLM, embedding, STT) into src/models/ with clear subdirectories. Extract credential schema and CRUD into src/provider/config/. Move logos into src/provider/. The llm/provider/ directory now only contains the SDK factory and model-listing service.

No HTTP route paths changed — only internal module boundaries.
