+++ packages/server/src/models/llm/ollama.ts
@@ -5,7 +5,7 @@ import { getDb } from '@/db/client.js';
 import { ollamaModels } from '@/db/schema/providers.js';
 import { err, ok } from '@/lib/service-result.js';
 import type { ServiceResult } from '@/lib/service-result.js';
-import { ModelSchema } from '@/llm/provider/models.js';
+import { ModelSchema } from '@/models/llm/registry.js';
 
 export type OllamaModel = typeof ollamaModels.$inferSelect;
