+++ packages/server/src/models/llm/registry.ts
