+++ packages/server/src/models/stt/schema.ts
