+++ packages/server/src/usage/cost.ts
@@ -1,4 +1,4 @@
-import * as Models from '@/llm/provider/models.js';
+import * as Models from '@/models/llm/registry.js';
 import type { LanguageModelUsage } from 'ai';
 
 const TOKENS_PER_MILLION = 1_000_000;
